+++ test_app/lib/new.dart
@@ -1,7 +1,6 @@
 import 'package:accessibility_features/accessibility_features.dart';
+import 'package:accessibility_features/accessibility_screen.dart';
 import 'package:accessibility_features/accessible_heading_text.dart';
-import 'package:accessibility_features/accessible_text.dart';
-import 'package:accessibility_features/accessiblity_image.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
@@ -14,36 +13,31 @@ class MyWidget extends StatelessWidget {
       builder: (context, accessibilityFeatures, _) {
         return Scaffold(
           appBar: AppBar(
-            title: const Text('Next Page'),
-          ),
-          body: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              AccessibilityImage(
-                image: AssetImage('assets/hello.png'),
-                width: 200,
-                height: 200,
-                fit: BoxFit.cover,
-                alignment: Alignment.center,
-                colorBlendMode: BlendMode.colorBurn,
-                filterQuality: FilterQuality.high,
-              ),
-              SizedBox(height: 20),
-              AccessibleHeadingText(
-                "Heading changed",
-                style: TextStyle(
-                  fontSize: accessibilityFeatures.currentFontSize,
-                  color: Colors.amber,
-                ),
-              ),
-              AccessibleText(
-                "Normal text",
-                style: TextStyle(
-                  fontSize: accessibilityFeatures.currentFontSize,
-                  color: accessibilityFeatures.textColor,
+            title: const Text('Accessibility Features'),
+            actions: [
+              Builder(
+                // Wrap the IconButton with Builder widget
+                builder: (context) => IconButton(
+                  icon: const Icon(Icons.settings),
+                  onPressed: () {
+                    // Use the context from the Builder
+                    Navigator.push(
+                      context,
+                      MaterialPageRoute(
+                        builder: (context) => const AccessibilityScreen(),
+                      ),
+                    );
+                  },
                 ),
               ),
             ],
+          ),
+          body: const Padding(
+            padding: EdgeInsets.all(20),
+            child: AccessibleHeadingText(
+              "Hello This is Accessibility...",
+              style: TextStyle(fontSize: 17),
+            ),
           ),
         );
       },
